requirements.txt
streamlit
langchain
langchain-core
langchain-openai
langchain-groq
langchain-google-genai
faiss-cpu
sentence-transformers
torch
python-dotenv
requests
